Move around nav/tap bar buttons and re-style tap bar
