--- src/components/Nav/Nav.css
@@ -1,48 +1,0 @@
-nav {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  background-color: plum;
-  height: 4em;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-  z-index: 5;
-}
-
-.nav-icon {
-  height: 25px;
-  width: 25px;
-}
-
-.nav-button {
-  background-color: inherit;
-  border: none;
-  padding: none;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-decoration: none;
-  color: black;
-  font-size: 1em;
-}
-
-.nav-button:disabled {
-  cursor: not-allowed;
-  color: grey;
-}
-
-.nav-icon[favorited] {
-  background-color: purple;
-  animation: fadeOut 3s;
-}
-
-@keyframes fadeOut {
-  0% {
-    opacity: 1;
-  }
-  100% {
-    opacity: 0;
-  }
-}
