+++ src/components/TapBar/TapBar.css
@@ -1,0 +1,48 @@
+.tap-bar {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  height: 5em;
+  width: 100%;
+  display: flex;
+}
+
+.dislike-bg {
+  background-color: red;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+  width: 50%;
+  z-index: 1;
+}
+
+.like-bg {
+  background-color: green;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 50%;
+  height: 100%;
+  z-index: 1;
+}
+
+.nav-button {
+  background-color: inherit;
+  border: none;
+  padding: none;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-decoration: none;
+  color: white;
+  font-size: 1em;
+}
+
+.fas {
+  font-size: 2em;
+}
+.nav-button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
